pyproject.toml
[project]
name = "synth-phi-data"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.74.0",
    "extract-msg>=0.55.0",
    "faker-healthcare-system>=0.1.24",
    "faker>=38.0.0",
    "openpyxl>=3.1.5",
    "pdfplumber>=0.11.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-pptx>=1.0.2",
    "pyyaml>=6.0",
    "reportlab>=4.4.5",
    "rich>=13.0.0",
    "typer>=0.20.0",
]

[project.scripts]
medforge = "src.cli:main"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.5",
]
